push_swap intmax ok

+++ push_swap/src/libft/ft_atoi.c
@@ -3,19 +3,19 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: rtissera <rtissera@student.42.fr>          +#+  +:+       +#+        */
+/*   By: raphael <raphael@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:56:01 by rtissera          #+#    #+#             */
-/*   Updated: 2023/03/14 14:36:22 by rtissera         ###   ########.fr       */
+/*   Updated: 2023/04/13 14:14:17 by raphael          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/push_swap.h"
 
-int	ft_atoi(const char *nptr)
+long int	ft_atoi(const char *nptr)
 {
-	int	res;
-	int	sig;
+	long int	res;
+	long int	sig;
 
 	res = 0;
 	sig = 1;
